#14 Feat: Navigation Bar 투명화 구현
