+++ smuing/src/constants/FOOTER_INFOS.ts
