+++ smuing/src/constants/PROJECTS_DATA.ts
