+++ smuing/src/types/types.ts
@@ -1,0 +1,17 @@
+export type project = {
+  id: number
+  name: string
+  description: string
+  theme: string[]
+  github: string | null
+  year: number
+  member: string[]
+  stack: string[]
+  img: string
+  release: string | null
+}
+
+export type scroll = {
+  x: number
+  y: number
+}
